refactor: Comment 를 추가해, Comment 가 갖는 기본적인 특성을 상속 받을 수 있도록 하였음 ex) AlbumComment 가 Comment 를 상속 받음
