+++ src/main/java/com/skhu/practice/entity/base/Review.java
@@ -1,0 +1,4 @@
+package com.skhu.practice.entity.base;
+
+public class Review {
+}
